+++ src/engine/main.c
@@ -636,9 +636,9 @@ print_banner(bool want_copyright, bool want_deps)
 			(al_version_id >> 16) & 0xFF, (al_version_id >> 8) & 0xFF,
 			(al_version_id & 0xFF) - 1);
 		printf("\n");
-		printf("  Allegro: v%-8s  libmng: v%s\n", al_version, mng_version_text());
-		printf("   Dyad.c: v%-8s    zlib: v%s\n", dyad_getVersion(), zlibVersion());
-		printf("  Duktape: %s\n", DUK_GIT_DESCRIBE);
+		printf("    Allegro: v%-8s   libmng: v%s\n", al_version, mng_version_text());
+		printf("     Dyad.c: v%-8s     zlib: v%s\n", dyad_getVersion(), zlibVersion());
+		printf("    Duktape: v%d.%d.%d\n", DUK_VERSION / 10000, DUK_VERSION / 100 % 100, DUK_VERSION % 100);
 		free(al_version);
 	}
 }